requirements.txt
aiohttp==3.7.3
alpha-vantage==2.3.1
async-timeout==3.0.1
attrs==20.3.0
Brotli==1.0.9
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.19.5
pandas==1.1.5
plotly==4.14.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
retrying==1.3.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
gunicorn==20.0.4
